Move bus_model flask routes to blueprint, add testing docs to readme
